core: changed price type to BigDecimal. tests: added order tests.

+++ src/main/java/pojo/Trade.java
@@ -2,8 +2,9 @@ package pojo;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public record Trade(UUID id, UUID orderSellId, UUID orderBuyId, Double price, Long quantity) {
+public record Trade(UUID id, UUID orderSellId, UUID orderBuyId, BigDecimal price, Long quantity) {
 }
